Refactoring and encapsulation of ContactInfo and Map element
